+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Add "Create new Mermaid file" command to command palette. This is particularly useful in Mobile mode where the "New Mermaid file" context menu is not very visible.
+
 ## 0.2.0 - 2026-01-17
 
 ### Fixed
 
-- Fix plugin ID in manifest (use `mermaid-view` instead of `obsidian-mermaid-view`)
+- Fix plugin ID in manifest (use `mermaid-view` instead of `obsidian-mermaid-view`).
 
 ## 0.1.0